+++ firestore-palm-chatbot/functions/__tests__/vertex.test.ts
@@ -12,38 +12,46 @@ process.env.FIRESTORE_EMULATOR_HOST = '127.0.0.1:8080';
 // // We mock out the config here instead of setting environment variables directly
 jest.mock('../src/config', () => ({
   default: {
-    collectionName: 'discussionsTest/{discussionId}/messages',
+    projectId: 'test-project',
+    collectionName: 'discussionsTestVertex/{discussionId}/messages',
     location: 'us-central1',
     orderField: 'createTime',
     promptField: 'prompt',
     responseField: 'response',
     enableDiscussionOptionOverrides: true,
-    candidatesField: 'candidates',
+    provider: 'vertex',
+    model: 'chat-bison@001',
   },
 }));
 
 // // mock to check the arguments passed to the annotateVideo function+
 const mockAPI = jest.fn();
+import {helpers} from '@google-cloud/aiplatform';
 
-// // Mock the video intelligence  clent
-jest.mock('@google-ai/generativelanguage', () => {
+jest.mock('@google-cloud/aiplatform', () => {
   return {
-    ...jest.requireActual('@google-ai/generativelanguage'),
-    DiscussServiceClient: function mockedClient() {
-      return {
-        generateMessage: async (args: unknown) => {
-          mockAPI(args);
-          return [
-            {
-              candidates: [
-                {
-                  content: 'test response',
-                },
-              ],
-            },
-          ];
-        },
-      };
+    ...jest.requireActual('@google-cloud/aiplatform'),
+    v1: {
+      PredictionServiceClient: function mockedClient() {
+        return {
+          predict: async (args: unknown) => {
+            mockAPI(args);
+            return [
+              {
+                predictions: [
+                  helpers.toValue({
+                    candidates: [
+                      {
+                        content: 'test response',
+                      },
+                    ],
+                  }),
+                ],
+              },
+            ];
+          },
+        };
+      },
     },
   };
 });
@@ -70,13 +78,18 @@ const wrappedGenerateMessage = fft.wrap(
 ) as WrappedFirebaseFunction;
 
 const firestoreObserver = jest.fn();
+let collectionName: string;
 
 describe('generateMessage', () => {
   let unsubscribe: (() => void) | undefined;
-  const collectionName = config.collectionName.replace('{discussionId}', '1');
 
   // clear firestore
   beforeEach(async () => {
+    const randomInteger = Math.floor(Math.random() * 1000000);
+    collectionName = config.collectionName.replace(
+      '{discussionId}',
+      randomInteger.toString()
+    );
     jest.clearAllMocks();
     await fetch(
       `http://${process.env.FIRESTORE_EMULATOR_HOST}/emulator/v1/projects/dev-extensions-testing/databases/(default)/documents`,
@@ -108,7 +121,7 @@ describe('generateMessage', () => {
 
     await simulateFunctionTriggered(wrappedGenerateMessage)(ref);
 
-    expectNoOp();
+    await expectNoOp();
   });
 
   test('should not run if the prompt field is empty', async () => {
@@ -123,7 +136,7 @@ describe('generateMessage', () => {
 
     await simulateFunctionTriggered(wrappedGenerateMessage)(ref);
 
-    expectNoOp();
+    await expectNoOp();
   });
 
   test('should not run if the prompt field is not a string', async () => {
@@ -138,19 +151,27 @@ describe('generateMessage', () => {
 
     await simulateFunctionTriggered(wrappedGenerateMessage)(ref);
 
-    expectNoOp();
+    await expectNoOp();
   });
 
   test('should not run if response field is set from the start', async () => {
     const message = {
       prompt: 'hello chat bison',
-      [config.responseField]: 'user set response for some reason',
+      [config.responseField]: 'user set response field for some reason',
     };
     const ref = await admin.firestore().collection(collectionName).add(message);
 
     await simulateFunctionTriggered(wrappedGenerateMessage)(ref);
 
-    expectNoOp();
+    const firestoreCallData = firestoreObserver.mock.calls.map(call => {
+      return call[0].docs[0].data();
+    });
+
+    expect(firestoreCallData.length).toBe(2);
+    expect(firestoreCallData[0]).toEqual({
+      prompt: 'hello chat bison',
+      response: 'user set response field for some reason',
+    });
   });
 
   test('should not run if status field is set from the start', async () => {
@@ -162,7 +183,16 @@ describe('generateMessage', () => {
 
     await simulateFunctionTriggered(wrappedGenerateMessage)(ref);
 
-    expectNoOp();
+    const firestoreCallData = firestoreObserver.mock.calls.map(call => {
+      return call[0].docs[0].data();
+    });
+
+    expect(firestoreCallData.length).toBe(2);
+    expect(firestoreCallData[0]).toEqual({
+      prompt: 'hello chat bison',
+      status: 'user set status field for some reason',
+    });
+    expectToHaveKeys(firestoreCallData[1], ['prompt', 'status', 'createTime']);
   });
 
   test("should update initial record with createTime if it doesn't have it", async () => {
@@ -247,23 +277,30 @@ describe('generateMessage', () => {
     });
     expect(firestoreCallData[2].response).toEqual('test response');
 
+    const prompt = {
+      messages: [
+        {
+          author: '0',
+          content: 'hello chat bison',
+        },
+      ],
+      context: '',
+      examples: [],
+    };
+
+    const instanceValue = helpers.toValue(prompt);
+    const instances = [instanceValue!];
+
+    const parameter = {};
+
+    const parameters = helpers.toValue(parameter);
+
     // verify SDK is called with expected arguments
     const expectedRequestData = {
-      candidateCount: undefined,
-      model: 'models/chat-bison-001',
-      prompt: {
-        messages: [
-          {
-            author: '0',
-            content: 'hello chat bison',
-          },
-        ],
-        context: undefined,
-        examples: [],
-      },
-      topP: undefined,
-      topK: undefined,
-      temperature: undefined,
+      endpoint:
+        'projects/test-project/locations/us-central1/publishers/google/models/chat-bison@001',
+      instances,
+      parameters,
     };
     // we expect the mock API to be called once
     expect(mockAPI).toHaveBeenCalledTimes(1);
@@ -337,23 +374,30 @@ describe('generateMessage', () => {
     });
     expect(firestoreCallData[3].response).toEqual('test response');
 
+    const prompt = {
+      messages: [
+        {
+          author: '0',
+          content: 'hello chat bison',
+        },
+      ],
+      context: '',
+      examples: [],
+    };
+
+    const instanceValue = helpers.toValue(prompt);
+    const instances = [instanceValue!];
+
+    const parameter = {};
+
+    const parameters = helpers.toValue(parameter);
+
     // verify SDK is called with expected arguments
     const expectedRequestData = {
-      candidateCount: undefined,
-      model: 'models/chat-bison-001',
-      prompt: {
-        messages: [
-          {
-            author: '0',
-            content: 'hello chat bison',
-          },
-        ],
-        context: undefined,
-        examples: [],
-      },
-      topP: undefined,
-      topK: undefined,
-      temperature: undefined,
+      endpoint:
+        'projects/test-project/locations/us-central1/publishers/google/models/chat-bison@001',
+      instances,
+      parameters,
     };
     // we expect the mock API to be called once
     expect(mockAPI).toHaveBeenCalledTimes(1);
@@ -367,14 +411,15 @@ const simulateFunctionTriggered =
     const data = (await ref.get()).data() as {[key: string]: unknown};
     const beforeFunctionExecution = fft.firestore.makeDocumentSnapshot(
       data,
-      `discussionsTest/1/messages/${ref.id}`
+      `${collectionName}/${ref.id}`
     ) as DocumentSnapshot;
     const change = fft.makeChange(before, beforeFunctionExecution);
     await wrappedFunction(change);
     return beforeFunctionExecution;
   };
 
-const expectNoOp = () => {
+const expectNoOp = async () => {
+  await new Promise(resolve => setTimeout(resolve, 500));
   expect(firestoreObserver).toHaveBeenCalledTimes(1);
   expect(mockAPI).toHaveBeenCalledTimes(0);
 };