+++ firestore-palm-chatbot/functions/src/config.ts
@@ -15,6 +15,9 @@
  */
 
 export interface Config {
+  location: string;
+  projectId: string;
+  instanceId: string;
   model: string;
   context?: string;
   promptField: string;
@@ -27,13 +30,37 @@ export interface Config {
   topK?: number;
   candidateCount?: number;
   candidatesField?: string;
+  provider: string;
+  apiKey?: string;
+}
+
+function getModel() {
+  switch (process.env.PALM_API_PROVIDER) {
+    case 'generative':
+      switch (process.env.MODEL) {
+        default:
+          return 'chat-bison-001';
+      }
+    default:
+      switch (process.env.MODEL) {
+        case 'codechat-bison':
+          return 'codechat-bison@001';
+        default:
+          return 'chat-bison@001';
+      }
+  }
 }
 
 const config: Config = {
+  // system defined
+  location: process.env.LOCATION!,
+  projectId: process.env.PROJECT_ID!,
+  instanceId: process.env.EXT_INSTANCE_ID!,
+  // user defined
   collectionName:
     process.env.COLLECTION_NAME ||
     'users/{uid}/discussions/{discussionId}/messages',
-  model: process.env.MODEL || 'models/chat-bison-001',
+  model: getModel(),
   context: process.env.CONTEXT,
   promptField: process.env.PROMPT_FIELD || 'prompt',
   responseField: process.env.RESPONSE_FIELD || 'response',
@@ -49,6 +76,8 @@ const config: Config = {
     ? parseInt(process.env.CANDIDATE_COUNT)
     : 1,
   candidatesField: process.env.CANDIDATES_FIELD || 'candidates',
+  provider: process.env.PALM_API_PROVIDER || 'vertex',
+  apiKey: process.env.API_KEY,
 };
 
 export default config;