+++ firestore-palm-summarize-text/functions/src/config.ts
@@ -15,13 +15,37 @@
  */
 
 export interface Config {
+  location: string;
+  projectId: string;
+  instanceId: string;
   collectionName: string;
   textField: string;
   responseField: string;
   targetSummaryLength?: number;
+  provider: string;
+  model: string;
+  apiKey?: string;
+}
+
+function getModel() {
+  switch (process.env.PALM_API_PROVIDER) {
+    case 'generative':
+      switch (process.env.MODEL) {
+        default:
+          return 'text-bison-001';
+      }
+    default:
+      switch (process.env.MODEL) {
+        default:
+          return 'text-bison@001';
+      }
+  }
 }
 
 const config: Config = {
+  location: process.env.LOCATION!,
+  projectId: process.env.PROJECT_ID!,
+  instanceId: process.env.EXT_INSTANCE_ID!,
   collectionName:
     process.env.COLLECTION_NAME || 'summaries/{summaryId}/messages',
   textField: process.env.TEXT_FIELD || 'text',
@@ -29,6 +53,9 @@ const config: Config = {
   targetSummaryLength: process.env.TARGET_SUMMARY_LENGTH
     ? parseInt(process.env.TARGET_SUMMARY_LENGTH)
     : undefined,
+  provider: process.env.PALM_API_PROVIDER || 'vertex',
+  model: getModel(),
+  apiKey: process.env.API_KEY,
 };
 
 export default config;